Compute min angle (#12)
* Use double precision

* Remove warnings for unused vars

* Fix formatting

* Compute the min angle for a vertex

* Add method for computing the min angle

+++ include/ctag/types.h
@@ -8,7 +8,7 @@ namespace ctag {
 
 class Types {
 public:
-    typedef float Coordinate;
+    typedef double Coordinate;
     typedef ctag::Point<Coordinate, 2> Point;
     typedef Point Direction;
